Modify files

+++ Java/Actividad3/ej4.java
@@ -6,8 +6,8 @@ public class ej4 {
         int empleadosBarna = 250;
         int totalEmpleados = empleadosBarna+empleadosMadrid;
 
-        float porcMadrid = (float)empleadosMadrid*100/totalEmpleados;
-        float porcBarna = (float)empleadosBarna*100/totalEmpleados;
+        double porcMadrid = (double)empleadosMadrid*100/totalEmpleados;
+        double porcBarna = (double)empleadosBarna*100/totalEmpleados;
 
         System.out.println("El porcentaje de empleados de Madrid es " + porcMadrid + " y el porcentaje de empleados en Barcelona es un " + porcBarna);
     }
